pyproject.toml
[project]
name = "discord-music-bot"
requires-python = ">=3.9"
version = "1.0.0"
dependencies = [
    "aiohttp==3.9.0",
    "discord.py==2.3.2",
    "PyNaCl==1.5.0",
    "python-dotenv==1.0.0",
    "validators==0.22.0",
    "yt-dlp==2023.11.16",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []